requirements.txt
Flask==2.3.3
Werkzeug==2.3.7

# AI/ML Dependencies for Proctoring
opencv-python==4.8.1.78
mediapipe==0.10.7
tensorflow==2.13.0
numpy==1.24.3
scipy==1.11.3
pillow==10.0.1

# Audio Processing for Whisper
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
soundfile==0.12.1
pyaudio==0.2.11

# WebRTC and Real-time Processing
aiortc==1.5.0
websockets==11.0.3
asyncio==3.4.3

# PDF Generation and Reports
reportlab==4.0.4
matplotlib==3.8.0
seaborn==0.13.0

# Database and Data Processing
pandas==2.1.1
sqlalchemy==2.0.21

# Additional Security and Utilities
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0